requirements.txt
alembic==1.10.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cors==1.0.1
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
psycopg2==2.9.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.5.post1
tldextract==3.4.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zope.event==4.6
zope.interface==5.5.2
